+++ src/main/java/com/company/finance/finance_manager/repository/InvoiceRepository.java
@@ -3,5 +3,5 @@ package com.company.finance.finance_manager.repository;
 import com.company.finance.finance_manager.entity.Invoice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvoiceRepository extends JpaRepository<Invoice, Long> {
+public interface InvoiceRepository extends JpaRepository<Invoice, Integer> {
 }
